DP-1169: Move ExpandableRow to components
