+++ src/components/Layout/DetailsSidebar.vue
